requirements.txt
django
psycopg2
django-cleanup
Pillow
django-admin-numeric-filter